requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
opencv-python>=4.8.0
pytorch-msssim>=1.0.0